git-svn-id: https://svn.vis.ethz.ch/svn/kob_gamelab09/trunk@506 fc16ba98-2866-4055-be8f-df951a5456ab

+++ src/ProjectMagma/ProjectMagma/Simulation/Properties/Player/PlayerControllerProperty.cs
@@ -113,9 +113,17 @@ namespace ProjectMagma.Simulation
         public void OnDetached(Entity player)
         {
             player.Update -= OnUpdate;
-            Game.Instance.Simulation.EntityManager.Remove(arrow);
-            if(flame != null)
+
+            if (arrow != null && Game.Instance.Simulation.EntityManager.ContainsEntity(arrow))
+            {
+                Game.Instance.Simulation.EntityManager.Remove(arrow);
+            }
+
+            if (flame != null && Game.Instance.Simulation.EntityManager.ContainsEntity(flame))
+            {
                 Game.Instance.Simulation.EntityManager.Remove(flame);
+            }
+
             Game.Instance.Simulation.EntityManager.EntityRemoved -= EntityRemovedHandler;
             if (player.HasProperty("collision"))
             {
